% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_app_daily_usage}
\alias{create_app_daily_usage}
\title{Daily app usage chart}
\usage{
create_app_daily_usage(apps_usage)
}
\arguments{
\item{apps_usage}{Second element returned by \link{create_app_ranking}.}
}
\value{
A calendar chart displaying daily app usage.
}
\description{
Leverages echarts4r.
See \url{https://echarts4r.john-coene.com/articles/chart_types.html#calendar-1}.
.
}
