% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_app_daily_usage_chart}
\alias{create_app_daily_usage_chart}
\title{Daily app usage chart}
\usage{
create_app_daily_usage_chart(app_usage, ...)
}
\arguments{
\item{app_usage}{Returned by \link{get_app_daily_usage}.}

\item{...}{Pass down any param to \link{create_calendar_chart}.}
}
\value{
A calendar chart displaying daily app usage.
}
\description{
Leverages echarts4r.
See \url{https://echarts4r.john-coene.com/articles/chart_types.html#calendar-1}.
.
}
