% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tables.R
\name{create_app_ranking_table}
\alias{create_app_ranking_table}
\title{Create app ranking table}
\usage{
create_app_ranking_table(ranking)
}
\arguments{
\item{ranking}{Ranked data from \link{create_app_ranking}.}
}
\value{
An htmlwidget table containing app ranking, owner sorted by
usage.
}
\description{
Leverages the toastui htmlwidget.
See \url{https://dreamrs.github.io/toastui/articles/extras/grid.html}.
}
