% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{create_apps_consumer_ranking}
\alias{create_apps_consumer_ranking}
\title{Create Shiny apps consumer ranking}
\usage{
create_apps_consumer_ranking(apps, users)
}
\arguments{
\item{apps}{Get from \link[connectapi]{get_usage_shiny}. Can be reactive or not.}

\item{users}{Get from \link[connectapi]{get_users}. Can be reactive or not.}
}
\value{
A 3 columns tibble with apps consumer sorted by number of view. The role
columns allows further analysis.
}
\description{
Sort consumers by number of views.
}
