% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_apps_consumer_ranking_chart}
\alias{create_apps_consumer_ranking_chart}
\title{Create apps consumer ranking bar chart}
\usage{
create_apps_consumer_ranking_chart(ranking)
}
\arguments{
\item{ranking}{Data obtained from \link{create_apps_consumer_ranking}.}
}
\value{
A bar chart with consumer sorted by descending number of total views.
}
\description{
Create apps consumer ranking bar chart
}
