% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_calendar_chart}
\alias{create_calendar_chart}
\title{Generic calendar chart generator}
\usage{
create_calendar_chart(data, max, range, title)
}
\arguments{
\item{data}{Calendar chart data.}

\item{max}{Maximum value.}

\item{range}{Date range.}

\item{title}{Chart title.}
}
\value{
An echarts4r calendar chart
}
\description{
Generic calendar chart generator
}
