% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_cumulated_duration_per_user}
\alias{create_cumulated_duration_per_user}
\title{Create cumulated app duration/user}
\usage{
create_cumulated_duration_per_user(
  apps_usage,
  start_date = NULL,
  end_date = NULL,
  selected_app
)
}
\arguments{
\item{apps_usage}{First element returned by \link{create_app_ranking}.
Can be reactive.}

\item{start_date}{Default to minimum calendar_data date. Could also be
an input value with Shiny.}

\item{end_date}{Default to maximum calendar_data date. Could also be
an input value with Shiny.}

\item{selected_app}{Selected app name (string). You'll need a selectInput for instance
wrapped by \link[shiny]{reactive}.}
}
\value{
An echarts4r barchart.
}
\description{
Bar chart
}
