% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_dev_project_overview}
\alias{create_dev_project_overview}
\title{Create developer project network overview}
\usage{
create_dev_project_overview(ranking, client, apps_usage)
}
\arguments{
\item{ranking}{Developer ranking. See \link{create_dev_ranking}. Useful
to get developers sorted by decreasing number of projects in the shiny selectInput.}

\item{client}{RSC client. See \link{create_rsc_client}.}

\item{apps_usage}{First element returned by \link{create_app_ranking}.}
}
\value{
A visNetwork htmlwidget with developer projects.
}
\description{
Leverages visNetwork.
}
