% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{create_dev_ranking}
\alias{create_dev_ranking}
\title{Create a developer ranking}
\usage{
create_dev_ranking(users, content)
}
\arguments{
\item{users}{Get from \link[connectapi]{get_users}.}

\item{content}{Get from \link[connectapi]{get_content}.}
}
\value{
A tibble with developer ranked by decreasing number of apps.
}
\description{
Rank developers by number of published apps
}
