% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_dev_ranking_chart}
\alias{create_dev_ranking_chart}
\title{Create developers ranking bar chart}
\usage{
create_dev_ranking_chart(ranking, threshold)
}
\arguments{
\item{ranking}{Obtained after calling \link{create_dev_ranking}.}

\item{threshold}{Minimum number of app threshold. You'll need a numericInput
wrapped by \link[shiny]{reactive}.}
}
\value{
An echarts4r bar chart.
}
\description{
Devs are ranked by number of developed apps.
See \link{create_dev_ranking}.
}
