% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_pie_chart}
\alias{create_pie_chart}
\title{Create standard pie chart}
\usage{
create_pie_chart(data)
}
\arguments{
\item{data}{Get from \link{sort_users_by_role}.}
}
\value{
A pie chart.
}
\description{
Pie chart with percentage data
}
