% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rsconnect.R
\name{create_rsc_client}
\alias{create_rsc_client}
\title{Create a connection to RStudio Connect server}
\usage{
create_rsc_client()
}
\description{
Leverages connectapi toolkit. Expect to have \code{Sys.getenv("CONNECT_SERVER")} and
\code{Sys.getenv("CONNECT_API_KEY")} properly setup.
}
