% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tables.R
\name{create_table_sparkline}
\alias{create_table_sparkline}
\title{Create sparkline graph for app usage table}
\usage{
create_table_sparkline(table)
}
\arguments{
\item{table}{Obtained after calling \link{generate_table}.}
}
\value{
Adds a sparkline bar chart to the existing table.
}
\description{
This function may be used by \link{generate_table} internally.
}
\keyword{internal}
