% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_user_daily_consumption_chart}
\alias{create_user_daily_consumption_chart}
\title{Daily app consumption for selected user}
\usage{
create_user_daily_consumption_chart(usage)
}
\arguments{
\item{usage}{Get from \link{get_user_daily_consumption}.}
}
\value{
An echarts4r calendar chart
}
\description{
Daily app consumption for selected user
}
