% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_app_daily_usage}
\alias{get_app_daily_usage}
\title{Daily app usage}
\usage{
get_app_daily_usage(apps_usage, selected_app)
}
\arguments{
\item{apps_usage}{Second element returned by \link{create_app_ranking}.}

\item{selected_app}{Selected app name (string). You'll need a selectInput for instance.
wrapped by \link[shiny]{reactive}.}
}
\value{
Calendar data for daily app usage.
}
\description{
Used by \link{create_app_daily_usage_chart}
.
}
