% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_max_rsc_apps_usage}
\alias{get_max_rsc_apps_usage}
\title{Get most used app}
\usage{
get_max_rsc_apps_usage(logs)
}
\arguments{
\item{logs}{Obtained after calling \code{get_rsc_apps_usage}.}
}
\value{
A numeric value.
}
\description{
Get most used app
}
