% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_rsc_developer_apps_count}
\alias{get_rsc_developer_apps_count}
\title{Get app count for each developer}
\usage{
get_rsc_developer_apps_count(developer, logs)
}
\arguments{
\item{developer}{Developer unique id.}

\item{logs}{All data logs.}
}
\value{
A number with all apps
}
\description{
Get app count for each developer
}
