% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_user_daily_consumption}
\alias{get_user_daily_consumption}
\title{Get daily shiny app usage for a given user}
\usage{
get_user_daily_consumption(content, users, apps, selected_user)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}. Can be reactive.}

\item{users}{Get from \link[connectapi]{get_users}. Can be reactive.}

\item{apps}{Get from \link[connectapi]{get_usage_shiny}. Can be reactive.}

\item{selected_user}{User to select.
You'll need a selectInput wrapped by \link[shiny]{reactive}.}
}
\value{
A list. [\link{1}] contains the row events filtered for the
given user. [\link{2}] is 2 columns tibble containing daily app consumption
for given user (grouped by dates).
}
\description{
Get daily shiny app usage for a given user
}
