% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{merge_rsc_data}
\alias{merge_rsc_data}
\title{Merge RStudio Connect data together}
\usage{
merge_rsc_data(content, users, apps)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}.}

\item{users}{Get from \link[connectapi]{get_users}.}

\item{apps}{Get from \link[connectapi]{get_usage_shiny}.}
}
\value{
A list of 3 tibbles. \verb{[[1]]}: light RSC content data (less columns);
\verb{[[2]]}: light RSC users data (less columns); \verb{[[3]]}: merged apps_usage data.
}
\description{
See \link{create_app_ranking}.
}
