% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{process_rsc_content}
\alias{process_rsc_content}
\title{Process RSC content}
\usage{
process_rsc_content(content)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}.}
}
\value{
A 3 columns tibble with content guid, app_name and owner guid.
Owner guid is the same returned in \link{process_rsc_user}.
}
\description{
Select relevant information for RSC content data. See
\link{merge_rsc_data}.
}
