% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{process_rsc_user}
\alias{process_rsc_user}
\title{Process RSC user}
\usage{
process_rsc_user(users)
}
\arguments{
\item{users}{Get from \link[connectapi]{get_users}.}
}
\value{
A 2 columns tibble with user guid and username.
}
\description{
Select relevant information for RSC users data. See
\link{create_app_ranking}.
}
