% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{sort_content_by_access}
\alias{sort_content_by_access}
\title{Sort RStudio Connect content by access type}
\usage{
sort_content_by_access(content)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}.}
}
\value{
A tibble with content grouped by access type.
}
\description{
Sort RStudio Connect content by access type
}
