% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{sort_content_by_pyversion}
\alias{sort_content_by_pyversion}
\title{Sort RStudio Connect content by python version}
\usage{
sort_content_by_pyversion(content)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}.}
}
\value{
A tibble with content grouped by python version.
}
\description{
Sort RStudio Connect content by python version
}
