% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{sort_users_by_role}
\alias{sort_users_by_role}
\title{Sort RStudio Connect users by role}
\usage{
sort_users_by_role(users)
}
\arguments{
\item{users}{Get from \link[connectapi]{get_users}.}
}
\value{
A tibble with user grouped by role.
}
\description{
Users are grouped by user_role to check the
server role repartition.
}
