% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_calendar_chart}
\alias{create_calendar_chart}
\title{Generic calendar chart generator}
\usage{
create_calendar_chart(
  calendar_data,
  title,
  start_date = min(calendar_data$Date),
  end_date = max(calendar_data$Date),
  callback = NULL
)
}
\arguments{
\item{calendar_data}{Calendar chart data.}

\item{title}{Chart title.}

\item{start_date}{Default to minimum calendar_data date. Could also be
an input value with Shiny.}

\item{end_date}{Default to maximum calendar_data date. Could also be
an input value with Shiny.}

\item{callback}{R function to pass \link[htmlwidgets]{onRender}. This is
useful to access the widget API on the R side at render time
and add events on the fly.}
}
\value{
An echarts4r calendar chart
}
\description{
Generic calendar chart generator
}
