% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-charts.R
\name{create_cumulated_hits_per_user}
\alias{create_cumulated_hits_per_user}
\title{Create cumulated app hits/user}
\usage{
create_cumulated_hits_per_user(apps_usage, input)
}
\arguments{
\item{apps_usage}{First element returned by \link{create_app_ranking}.}

\item{input}{Shiny input object.}
}
\value{
An echarts4r barchart.
}
\description{
Bar chart
}
