% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tables.R
\name{generate_table}
\alias{generate_table}
\title{Generate htmlWidgets table}
\usage{
generate_table(logs, sparkline = FALSE)
}
\arguments{
\item{logs}{Obtained with \code{get_rsc_apps_usage}.}

\item{sparkline}{Whether to draw a sparkline.}
}
\value{
An htmlwidget table.
}
\description{
Leverage toastui API.
}
