% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_rsc_app_dates}
\alias{get_rsc_app_dates}
\title{Extract app usage dates}
\usage{
get_rsc_app_dates(app, logs)
}
\arguments{
\item{app}{app_name to filter.}

\item{logs}{Given by RSC database.}
}
\value{
A tibble containing dates of usage for
the given app as well as the correponding session duration.
}
\description{
Extract app usage dates
}
