% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_rsc_app_dates}
\alias{get_rsc_app_dates}
\title{Extract app usage dates}
\usage{
get_rsc_app_dates(logs, app)
}
\arguments{
\item{logs}{Given by RSC database.}

\item{app}{app_name to filter.}
}
\value{
A vector containing dates of usage for
the given app.
}
\description{
Extract app usage dates
}
