% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_rsc_developer_apps_list}
\alias{get_rsc_developer_apps_list}
\title{Get app list for each developer + usage}
\usage{
get_rsc_developer_apps_list(logs, developer)
}
\arguments{
\item{logs}{RSC logs.}

\item{developer}{Unique user id. Typically
app_developer_id key.}
}
\value{
Tibble containing app name + count
}
\description{
Get app list for each developer + usage
}
