% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{sort_content_by_access}
\alias{sort_content_by_access}
\title{Sort RStudio Connect content by access type}
\usage{
sort_content_by_access(content, start_date = NULL, end_date = NULL)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}.}

\item{start_date}{Date filter.}

\item{end_date}{Date filter.}
}
\value{
A tibble with content grouped by access type.
}
\description{
Sort RStudio Connect content by access type
}
