% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{sort_content_by_rversion}
\alias{sort_content_by_rversion}
\title{Sort RStudio Connect content by R version}
\usage{
sort_content_by_rversion(content)
}
\arguments{
\item{content}{Get from \link[connectapi]{get_content}.}
}
\value{
A tibble with content grouped by R version.
}
\description{
Sort RStudio Connect content by R version
}
